giant rework in progress
everything's now in es6, which is nice.

everything's being shrunk, which is nice.

i'm removing some 'features' that really just don't belong in this app.

that might not be nice, but it really needs to be done.

next up: figuring out which livereload sticks the least amount of stuff
in whatever it is we're trying to serve.

also, nixing as many dependencies as possible, because this tool does
_not_ need to be this large.